Fix disappearing symbol bug.
Instead of querying for all the rows where is_current is 1 to get all
the symbols in the list, make a `distinct symbol` projection query.

It's apparently possible that is_current could be 0, so any added
symbols will no longer appear in the list, but will remain in the database.